Split fxpay into modules (part of bug 1100618)
