+++ README.rst
@@ -34,6 +34,21 @@ This is pretty standard for any Node project so you you might already have it.
 To test that you have your path set up, type ``which grunt`` and make
 sure you see a path to the executable.
 
+Compression
+~~~~~~~~~~~
+
+To build yourself a compressed version of ``fxpay.js``, run this::
+
+    grunt compress
+
+The compressed source file will appear in the ``build`` directory
+as ``fxpay.min.js``.
+
+**IMPORTANT**: To use this library in a web page you have to
+compress it first because the source code spans multiple files.
+The usage instructions above explain how to install public releases from
+Bower which is of course easier.
+
 Running Tests
 ~~~~~~~~~~~~~
 
@@ -68,28 +83,34 @@ To check for syntax errors (lint), run::
 
     grunt jshint
 
-Compression
-~~~~~~~~~~~
+Create A Release
+~~~~~~~~~~~~~~~~
 
-To build yourself a compressed version of ``fxpay.js``, run this::
+You have to do a couple things to create a release:
 
-    grunt compress
+* Run ``grunt compress`` and add ``build/fxpay.min.js`` to ``dist``
+* Make sure the Changelog is up to date.
+  You'll probably just need to replace *unreleased* with the current date
+  for the current version.
+* Commit and push your changes.
+* Add and push a git tag corresponding to the version number so that bower
+  picks up the file. For example, tag a ``0.0.1`` release like ``git tag 0.0.1``
+* Bump the version for the next release. Library version numbers are
+  managed in multiple files.
+  To increment the version number and update all files at once,
+  run ``grunt bump``. In the Changelog, mark this new release number
+  as *unreleased*.
+* Commit and push your changes.
 
-The compressed source file will appear in the ``build`` directory.
-
-Bump Version
-~~~~~~~~~~~~
-
-Library version numbers are managed in multiple files. To increment
-the version number and update all files at once, run this::
-
-    grunt bump
-
-Commit the changes, tag, and push. For example, tag a ``0.0.1``
-release like ``git tag 0.0.1``.
 
 Changelog
 =========
+
+**0.0.5** (unreleased)
+
+* Split fxpay.js into smaller modules. This means you *must* minify the
+  source before you can use it. That is, unless you install the library
+  with Bower.
 
 **0.0.4** (2014-09-17)
