+++ example/index.html
@@ -35,7 +35,7 @@
       </div>
     </div>
     <script type="text/javascript" src="js/lib/jquery-1.9.1.min.js"></script>
-    <script type="text/javascript" src="fxpay.js"></script>
+    <script type="text/javascript" src="fxpay.min.js"></script>
     <script type="text/javascript" src="js/index.js"></script>
   </body>
 </html>
