+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "fxpay",
   "description": "Adds Firefox Marketplace payments to a web application",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "homepage": "https://github.com/mozilla/fxpay",
   "repository": {
     "type": "git",
