+++ tasks/packager.js
@@ -8,6 +8,7 @@ var grunt;
 exports.createTask = function(_grunt, siteDir) {
   grunt = _grunt;
   return function() {
+    grunt.task.run('compress');
     run(this.async(), siteDir);
   };
 };
@@ -19,11 +20,11 @@ function run(done, siteDir) {
   var pkgDir = buildDir + '/application';
   var pkgFile = pkgDir + '.zip';
 
-  shell('rm', ['-fr', pkgDir + '*'], function() {
+  shell('rm', ['-fr', (pkgDir || '/__thank-me-later__') + '*'], function() {
     shell('cp', ['-R', siteDir + '/', pkgDir], function() {
-      // Copy fxpay.js into the package root.
-      shell('cp', [workDir + '/lib/fxpay.js',
-                   pkgDir + '/fxpay.js'], function() {
+      // Copy minified fxpay.js into the package root.
+      shell('cp', [workDir + '/build/fxpay.min.js',
+                   pkgDir + '/fxpay.min.js'], function() {
         process.chdir(pkgDir);
         shell('zip', ['-Xr', pkgFile, './*'], function() {
           grunt.log.writeln('Package folder: ' + pkgDir);
